refactor: mergeRefs into useMergeRegs hook
